+++ src/jaydee/crawler.py
@@ -1,10 +1,10 @@
-from .scraper import Scraper, ScraperRule, ScraperOptions
-from . import utils
+from . import Scraper, ScraperRule, ScraperOptions, utils
 
 import logging
 from dataclasses import dataclass
 from datetime import datetime
 
+import playwright.async_api as pwa
 from playwright.async_api import async_playwright
 
 # Setup the scraper specific logger
@@ -18,11 +18,14 @@ class CrawlerOptions:
     # Whether or not the instance of Playwright will be headless.
     _headless: bool
 
-    # If not empty, the crawler wait for selector be present before parsing HTML.
-    _waitForSelector: str
+    # If not empty, the crawler waits for selector be present before parsing HTML.
+    _waitForSelector: str | None
 
     # Whether or not to wait for the network to be idle for half a second before parsing HTML.
     _waitForIdle: bool
+
+    # If not empty, the crawler waits for equivalent text to be visible on the page.
+    _waitForText: str | None
 
     # User agent to use for crawling
     _userAgents: list[str]
@@ -38,12 +41,14 @@ class CrawlerOptions:
         headless=True,
         waitForSelector=None,
         waitForIdle=False,
+        waitForText=None,
         scraperOptions=ScraperOptions(True),
         strict=True,
     ):
         """Setup default values."""
         self._headless = headless
         self._waitForSelector = waitForSelector
+        self._waitForText = waitForText
         self._waitForIdle = waitForIdle
         self._scraperOptions = scraperOptions
         self._strict = strict
@@ -67,7 +72,7 @@ class Crawler:
     def __init__(
         self,
         initial_url: str,
-        callback,
+        callback=None,
         rule: ScraperRule = None,
         child_of=None,
         options: CrawlerOptions = CrawlerOptions(),
@@ -120,7 +125,7 @@ class Crawler:
         The crawler runs until it's URL queue is empty and yields links of interest. When the URL queue is empty, Crawler
         invokes it's callback function `on_proceed` which should include any possible additions to the URL queue.
 
-        yields a list of urls whenever the crawler has succesfully scraped a list of links.
+        yields a list of urls whenever the crawler has successfully scraped a list of links.
         """
 
         # Start running
@@ -135,12 +140,16 @@ class Crawler:
                 return None
 
             page = await browser.new_page()
+            # Trick for attempting to bypass restrictions
+            await page.add_init_script(
+                "delete Object.getPrototypeOf(navigator).webdriver"
+            )
             response = await page.goto(url)
 
             logger.info(f"{url} retrieved with response: {response.status}")
 
             metadata = {
-                "time": datetime.today().strftime("%Y-%m-%d %H:%M:%S"),
+                "date": datetime.today().strftime("%Y-%m-%d %H:%M:%S"),
                 "url": url,
                 "base_url": utils.parse_base_url(url),
                 "domain": utils.parse_domain(url),
@@ -156,11 +165,15 @@ class Crawler:
 
             # If wait for idle is enabled, wait for network to be idle for half a second.
             # Takes precedence over waiting for a selector.
-            if self.options._waitForIdle is not None:
+            if self.options._waitForIdle:
                 await page.wait_for_load_state("networkidle")
             # Alternatively a selector can be used.
             elif self.options._waitForSelector is not None:
                 await page.wait_for_selector(self.options._waitForSelector)
+            elif self.options._waitForText is not None:
+                await pwa.expect(
+                    page.get_by_text(self.options._waitForText)
+                ).to_be_visible()
 
             html = await page.content()
             await page.close()
@@ -168,7 +181,9 @@ class Crawler:
             return {"doc": html, "metadata": metadata}
 
         async with async_playwright() as pw:
-            browser = await pw.chromium.launch(headless=self.options._headless)
+            browser = await pw.chromium.launch(
+                headless=self.options._headless, args=utils.get_chrome_arguments()
+            )
 
             while self.url_queue and self.running:
                 self.current_result = {}