+++ src/jaydee/webscraper.py
@@ -3,8 +3,7 @@ import logging
 from copy import deepcopy
 from dataclasses import dataclass
 
-from .scraper import Scraper
-from . import utils
+from . import Scraper, utils
 
 from playwright.async_api import async_playwright
 
@@ -25,17 +24,27 @@ class WebScraperOptions:
     # Max amount of tasks that can be run concurrently.
     _max_concurrent_tasks: int
 
+    # If not empty, the crawler wait for selector be present before parsing HTML.
+    _waitForSelector: str
+
+    # Whether or not to wait for the network to be idle for half a second before parsing HTML.
+    _waitForIdle: bool
+
     def __init__(
         self,
-        timeout: int = 2,
-        retries: int = 5,
+        timeout: int = 5,
+        retries: int = 3,
         pool_size: int = 3,
         max_concurrent_tasks: int = 8,
+        waitForSelector=None,
+        waitForIdle=False,
     ):
         self._timeout = timeout
         self._retries = retries
         self._pool_size = pool_size
         self._max_concurrent_tasks = max_concurrent_tasks
+        self._waitForSelector = waitForSelector
+        self._waitForIdle = waitForIdle
 
 
 class WebScraper:
@@ -49,26 +58,31 @@ class WebScraper:
         urls: list[str] = [],
         options: WebScraperOptions = WebScraperOptions(),
     ):
-        self.url_queue = urls
+        self.url_queue = []
+        self.add_urls(urls)
+
         self.scraper = scraper
         self.options = options
-        self.current_result = {}
+
+        self._current_result = {}
+        self._total_success = 0
+        self._total_failures = 0
 
     async def scrape_pages(self):
         """
-        Starts the web scraping coroutine.
+        Starts the page scraping coroutine.
         """
-        self.current_result = {
-            "results": [],
-            "success": 0,
-            "failures": 0,
-        }
-
         if not self.url_queue:
             logger.error("No URLs in queue, unable to web scrape.")
             return
 
         async with async_playwright() as pw:
+            self._current_result = {
+                "results": [],
+                "success": 0,
+                "failures": 0,
+            }
+
             browser = await pw.chromium.launch()
             contexts = [
                 await browser.new_context() for _ in range(self.options._pool_size)
@@ -98,9 +112,15 @@ class WebScraper:
                     )
 
                 await asyncio.gather(*tasks)
+
+                self.total_success += self.current_result["success"]
+                self.total_failures += self.current_result["failures"]
+
                 return self.current_result
             except Exception as e:
-                logger.error("Error in the webscraper start coroutine:")
+                logger.error(
+                    "Error occurred in the webscraper page scraping coroutine:"
+                )
                 logger.error(e)
             finally:
                 for context in contexts:
@@ -119,11 +139,16 @@ class WebScraper:
         Args:
             context: browser context provided by Playwright.
             url: URL of the webpage to scrape.
+            scraper: instance of a scraper to scrape the page with.
         """
         page = await context.new_page()
+        # Trick for attempting to bypass restrictions
+        await page.add_init_script("delete Object.getPrototypeOf(navigator).webdriver")
         try:
             logger.info(f"Scraping {url}...")
+
             await page.goto(url, timeout=self.options._timeout * 1000)
+            await self.__wait_for(page)
             content = await page.content()
 
             result = scraper.scrape(content)
@@ -137,7 +162,6 @@ class WebScraper:
             logger.error(e)
         finally:
             await page.close()
-            return {}
 
     async def scrape_page(self, url: str):
         """
@@ -146,12 +170,18 @@ class WebScraper:
         In case of error, returns an empty object.
         """
         async with async_playwright() as pw:
-            browser = await pw.chromium.launch(headless=False)
+            browser = await pw.chromium.launch(
+                headless=True, args=utils.get_chrome_arguments()
+            )
             context = await browser.new_context(
                 user_agent=utils.get_random_user_agent(),
                 viewport={"width": 1920, "height": 1080},
             )
             page = await context.new_page()
+            # Trick for attempting to bypass restrictions
+            await page.add_init_script(
+                "delete Object.getPrototypeOf(navigator).webdriver"
+            )
             try:
                 if not utils.validate_url(url):
                     logger.warning(
@@ -160,17 +190,58 @@ class WebScraper:
                     return {}
 
                 await page.goto(url, timeout=self.options._timeout * 1000)
-                html = await page.content()
+                await self.__wait_for(page)
 
+                html = await page.content()
                 result = self.scraper.scrape(html)
+                result["_content"] = html
+
                 return result
             except Exception as e:
                 logger.error(f"Error with scraping url: {url}")
                 logger.error(e)
             finally:
+                await page.close()
                 await browser.close()
-                return {}
 
     def add_urls(self, urls: list[str]):
-        """Adds urls to the list to be scraped."""
-        self.url_queue += urls
+        """Adds urls to the list to be scraped. URLs are validated before they are appended."""
+        for url in urls:
+            if not utils.validate_url(url):
+                logger.info(
+                    f"Attempting to add invalid URL: {url} to queue, skipping.."
+                )
+                continue
+
+            self.url_queue.append(url)
+
+    async def __wait_for(self, page):
+        """Wait for idle / selector."""
+        if self.options._waitForIdle:
+            await page.wait_for_load_state("networkidle")
+        elif self.options._waitForSelector is not None:
+            await page.wait_for_selector(self.options._waitForSelector)
+
+    @property
+    def current_result(self):
+        return self._current_result
+
+    @current_result.setter
+    def current_result(self, val):
+        self.current_result = val
+
+    @property
+    def total_success(self):
+        return self._total_success
+
+    @total_success.setter
+    def total_success(self, val):
+        self._total_success = val
+
+    @property
+    def total_failures(self):
+        return self._total_failures
+
+    @total_failures.setter
+    def total_failures(self, val):
+        self._total_failures = val